Handle missed checkout logic

+++ src/main/java/com/abdul/checkinservice/config/AppProperties.java
@@ -21,7 +21,7 @@ public class AppProperties {
 
     @Data
     public static class Shift {
-        private Float maxDuration;
+        private Integer maxDuration;
     }
 
     @Data
